Add files via upload
Program to generate Jacknife light curve data for an object and run Lombe-Scargle analysis to find the most likely period for each lightcurve and produce a histogram of these. The MCMC fits a Sinusoid to the light curve of an object using flat priors on A, p, phi, mu, nu. Emcee is used as the Markov Chain Monte Carlo Analysis. The best fitting period and SNR ratio are outputted to a text file.