Add files via upload
This program does a variety of tasks, taking as its input the ID_Values in a text folder and the results from the tsunami process. It allows us to fit a prediction for the CARMA and CAR(1) models to the lightcurves, plot a histogram for the nu values of the objects. This may be done for a specific number of objects, here we do it for the objects with Osc. Stat > 0.5. It also, plots the binned lightcurve, residuals and the lightcurve folded onto the period.