Add files via upload
This program fits a CARMA model and uses nested sampling to produce the evidence and posterior samples of this fit. The posterior and postsamples are saved. Other outputs are the lightcurve of the data, a distribution of the quality factors, a map of all of the autoregressive roots from the post samples and a textfile of useful values including evidence, period with errors and number of posterior samples within reasonable range.